# docker_images/csharp/wrapper/src/wwwroot/swagger-original.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "REST API definition for End-to-end testing of the Azure IoT SDKs.  All SDK APIs that are tested by our E2E tests need to be defined in this file.  This file takes some liberties with the API definitions.  In particular, response schemas are undefined, and error responses are also undefined.",
    "version" : "1.0.0",
    "title" : "Azure IOT End-to-End Test Wrapper Rest Api"
  },
...
  "schemes" : [ "https" ],
  "produces" : [ "application/json" ],
  "paths" : {
...
    "/module/connect/{transportType}" : {
      "put" : {
        "tags" : [ "module" ],
        "summary" : "Connect to the azure IoT Hub as a module",
        "operationId" : "Module_Connect",
        "consumes" : [ "application/json" ],
        "parameters" : [ {
          "name" : "transportType",
          "in" : "path",
          "description" : "Transport to use",
          "required" : true,
          "type" : "string",
          "enum" : [ "amqp", "amqpws", "mqtt", "mqttws", "http" ]
        }, {
          "name" : "connectionString",
          "in" : "query",
          "description" : "connection string",
          "required" : true,
          "type" : "string"
        }, {
          "in" : "body",
          "name" : "caCertificate",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/certificate"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "schema" : {
              "$ref" : "#/definitions/connectResponse"
            }
          }
        }
      }
    },
    "/module/{connectionId}/disconnect" : {
      "put" : {
        "tags" : [ "module" ],
        "summary" : "Disconnect the module",
        "operationId" : "Module_Disconnect",
        "parameters" : [ {
          "name" : "connectionId",
          "in" : "path",
          "description" : "Id for the connection",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    },
    "/module/connectFromEnvironment/{transportType}" : {
      "put" : {
        "tags" : [ "module" ],
        "summary" : "Connect to the azure IoT Hub as a module using the environment variables",
        "operationId" : "Module_ConnectFromEnvironment",
        "parameters" : [ {
          "name" : "transportType",
          "in" : "path",
          "description" : "Transport to use",
          "required" : true,
          "type" : "string",
          "enum" : [ "amqp", "amqpws", "mqtt", "mqttws", "http" ]
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "schema" : {
              "$ref" : "#/definitions/connectResponse"
            }
          }
        }
      }
    },
    "/module/createFromConnectionstring/{transportType}" : {
      "put" : {
        "tags" : [ "module" ],
        "summary" : "Create a module client from a connection string",
        "operationId" : "Module_CreateFromConnectionString",
        "consumes" : [ "application/json" ],
        "parameters" : [ {
          "name" : "transportType",
          "in" : "path",
          "description" : "Transport to use",
          "required" : true,
          "type" : "string",
          "enum" : [ "amqp", "amqpws", "mqtt", "mqttws", "http" ]
        }, {
          "name" : "connectionString",
          "in" : "query",
          "description" : "connection string",
          "required" : true,
          "type" : "string"
        }, {
          "in" : "body",
          "name" : "caCertificate",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/certificate"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "schema" : {
              "$ref" : "#/definitions/connectResponse"
            }
          }
        }
      }
    },
    "/module/createFromEnvironment/{transportType}" : {
      "put" : {
        "tags" : [ "module" ],
        "summary" : "Create a module client using the EdgeHub environment",
        "operationId" : "Module_CreateFromEnvironment",
        "parameters" : [ {
          "name" : "transportType",
          "in" : "path",
          "description" : "Transport to use",
          "required" : true,
          "type" : "string",
          "enum" : [ "amqp", "amqpws", "mqtt", "mqttws", "http" ]
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "schema" : {
              "$ref" : "#/definitions/connectResponse"
            }
          }
        }
      }
    },
    "/module/createFromX509/{transportType}" : {
      "put" : {
        "tags" : [ "module" ],
        "summary" : "Create a module client from X509 credentials",
        "operationId" : "Module_CreateFromX509",
        "parameters" : [ {
          "name" : "transportType",
          "in" : "path",
          "description" : "Transport to use",
          "required" : true,
          "type" : "string",
          "enum" : [ "amqp", "amqpws", "mqtt", "mqttws", "http" ]
        }, {
          "in" : "body",
          "name" : "X509",
          "required" : true,
          "schema" : {
            "type" : "object"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "schema" : {
              "$ref" : "#/definitions/connectResponse"
            }
          }
        }
      }
    },
    "/module/{connectionId}/connect2" : {
      "put" : {
        "tags" : [ "module" ],
        "summary" : "Connect the module",
        "operationId" : "Module_Connect2",
        "parameters" : [ {
          "name" : "connectionId",
          "in" : "path",
          "description" : "Id for the connection",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    },
    "/module/{connectionId}/reconnect" : {
      "put" : {
        "tags" : [ "module" ],
        "summary" : "Reconnect the module",
        "operationId" : "Module_Reconnect",
        "parameters" : [ {
          "name" : "connectionId",
          "in" : "path",
          "description" : "Id for the connection",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "forceRenewPassword",
          "in" : "query",
          "description" : "True to force SAS renewal",
          "required" : false,
          "type" : "boolean"
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    },
    "/module/{connectionId}/disconnect2" : {
      "put" : {
        "tags" : [ "module" ],
        "summary" : "Disonnect the module",
        "operationId" : "Module_Disconnect2",
        "parameters" : [ {
          "name" : "connectionId",
          "in" : "path",
          "description" : "Id for the connection",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    },
    "/module/{connectionId}/destroy" : {
      "put" : {
        "tags" : [ "module" ],
        "summary" : "Disonnect and destroy the module client",
        "operationId" : "Module_Destroy",
        "parameters" : [ {
          "name" : "connectionId",
          "in" : "path",
          "description" : "Id for the connection",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    },
    "/module/{connectionId}/enableTwin" : {
      "put" : {
        "tags" : [ "module" ],
        "summary" : "Enable module twins",
        "operationId" : "Module_EnableTwin",
        "parameters" : [ {
          "name" : "connectionId",
          "in" : "path",
          "description" : "Id for the connection",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    },
    "/module/{connectionId}/enableMethods" : {
      "put" : {
        "tags" : [ "module" ],
        "summary" : "Enable methods",
        "operationId" : "Module_EnableMethods",
        "parameters" : [ {
          "name" : "connectionId",
          "in" : "path",
          "description" : "Id for the connection",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    },
    "/module/{connectionId}/enableInputMessages" : {
      "put" : {
        "tags" : [ "module" ],
        "summary" : "Enable input messages",
        "operationId" : "Module_EnableInputMessages",
        "parameters" : [ {
          "name" : "connectionId",
          "in" : "path",
          "description" : "Id for the connection",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    },
    "/module/{connectionId}/twin" : {
      "get" : {
        "tags" : [ "module" ],
        "summary" : "Get the device twin",
        "operationId" : "Module_GetTwin",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "connectionId",
          "in" : "path",
          "description" : "Id for the connection",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "schema" : {
              "$ref" : "#/definitions/twin"
            }
          }
        }
      },
      "patch" : {
        "tags" : [ "module" ],
        "summary" : "Updates the device twin",
        "operationId" : "Module_PatchTwin",
        "consumes" : [ "application/json" ],
        "parameters" : [ {
          "name" : "connectionId",
          "in" : "path",
          "description" : "Id for the connection",
          "required" : true,
          "type" : "string"
        }, {
          "in" : "body",
          "name" : "twin",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/twin"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    },
    "/module/{connectionId}/twinDesiredPropPatch" : {
      "get" : {
        "tags" : [ "module" ],
        "summary" : "Wait for the next desired property patch",
        "operationId" : "Module_WaitForDesiredPropertiesPatch",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "connectionId",
          "in" : "path",
          "description" : "Id for the connection",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "schema" : {
              "$ref" : "#/definitions/twin"
            }
          }
        }
      }
    },
    "/module/{connectionId}/event" : {
      "put" : {
        "tags" : [ "module" ],
        "summary" : "Send an event",
        "operationId" : "Module_SendEvent",
        "consumes" : [ "application/json" ],
        "parameters" : [ {
          "name" : "connectionId",
          "in" : "path",
          "description" : "Id for the connection",
          "required" : true,
          "type" : "string"
        }, {
          "in" : "body",
          "name" : "eventBody",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/eventBody"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    },
    "/module/{connectionId}/outputEvent/{outputName}" : {
      "put" : {
        "tags" : [ "module" ],
        "summary" : "Send an event to a module output",
        "operationId" : "Module_SendOutputEvent",
        "consumes" : [ "application/json" ],
        "parameters" : [ {
          "name" : "connectionId",
          "in" : "path",
          "description" : "Id for the connection",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "outputName",
          "in" : "path",
          "required" : true,
          "type" : "string"
        }, {
          "in" : "body",
          "name" : "eventBody",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/eventBody"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    },
    "/module/{connectionId}/inputMessage/{inputName}" : {
      "get" : {
        "tags" : [ "module" ],
        "summary" : "Wait for a message on a module input",
        "operationId" : "Module_WaitForInputMessage",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "connectionId",
          "in" : "path",
          "description" : "Id for the connection",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "inputName",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "schema" : {
              "$ref" : "#/definitions/eventBody"
            }
          }
        }
      }
    },
    "/module/{connectionId}/waitForMethodAndReturnResponse/{methodName}" : {
      "put" : {
        "tags" : [ "module" ],
        "summary" : "Wait for a method call, verify the request, and return the response.",
        "description" : "This is a workaround to deal with SDKs that only have method call operations that are sync.  This function responds to the method with the payload of this function, and then returns the method parameters.  Real-world implemenatations would never do this, but this is the only same way to write our test code right now (because the method handlers for C, Java, and probably Python all return the method response instead of supporting an async method call)",
        "operationId" : "Module_WaitForMethodAndReturnResponse",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "connectionId",
          "in" : "path",
          "description" : "Id for the connection",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "methodName",
          "in" : "path",
          "description" : "name of the method to handle",
          "required" : true,
          "type" : "string"
        }, {
          "in" : "body",
          "name" : "requestAndResponse",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/methodRequestAndResponse"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    },
    "/module/{connectionId}/moduleMethod/{deviceId}/{moduleId}" : {
      "put" : {
        "tags" : [ "module" ],
        "summary" : "call the given method on the given module",
        "operationId" : "Module_InvokeModuleMethod",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "connectionId",
          "in" : "path",
          "description" : "Id for the connection",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "deviceId",
          "in" : "path",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "moduleId",
          "in" : "path",
          "required" : true,
          "type" : "string"
        }, {
          "in" : "body",
          "name" : "methodInvokeParameters",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/methodInvoke"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "schema" : {
              "type" : "object"
            }
          }
        }
      }
    },
    "/module/{connectionId}/deviceMethod/{deviceId}" : {
      "put" : {
        "tags" : [ "module" ],
        "summary" : "call the given method on the given device",
        "operationId" : "Module_InvokeDeviceMethod",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "connectionId",
          "in" : "path",
          "description" : "Id for the connection",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "deviceId",
          "in" : "path",
          "required" : true,
          "type" : "string"
        }, {
          "in" : "body",
          "name" : "methodInvokeParameters",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/methodInvoke"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "schema" : {
              "type" : "object"
            }
          }
        }
      }
    },
    "/module/{connectionId}/connectionStatus" : {
      "get" : {
        "tags" : [ "module" ],
        "summary" : "get the current connection status",
        "operationId" : "Module_GetConnectionStatus",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "connectionId",
          "in" : "path",
          "description" : "Id for the connection",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "schema" : {
              "type" : "string"
            }
          }
        }
      }
    },
    "/module/{connectionId}/connectionStatusChange" : {
      "get" : {
        "tags" : [ "module" ],
        "summary" : "wait for the current connection status to change and return the changed status",
        "operationId" : "Module_WaitForConnectionStatusChange",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "connectionId",
          "in" : "path",
          "description" : "Id for the connection",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "schema" : {
              "type" : "string"
            }
          }
        }
      }
    },
...
  },
  "definitions" : {
...
    "connectResponse" : {
      "type" : "object",
      "properties" : {
        "connectionId" : {
          "type" : "string",
          "description" : "unique identifier for this connection"
        }
      },
      "description" : "result of a connection to a service, device, or module client",
      "example" : {
        "connectionId" : "connectionId"
      }
    },
    "twin" : {
      "type" : "object",
      "properties" : {
        "desired" : {
          "type" : "object",
          "description" : "desired properties",
          "properties" : { }
        },
        "reported" : {
          "type" : "object",
          "description" : "reported properties",
          "properties" : { }
        }
      },
      "description" : "device twin or module twin",
      "example" : {
        "desired" : "{}",
        "reported" : "{}"
      }
    },
    "certificate" : {
      "type" : "object",
      "properties" : {
        "cert" : {
          "type" : "string",
          "description" : "unique identifier for this connection"
        }
      },
      "description" : "certificate in the body of a message"
    },
    "eventBody" : {
      "type" : "object",
      "properties" : {
        "body" : {
          "type" : "object",
          "description" : "payload to send to the method",
          "properties" : { }
        },
        "horton_flags" : {
          "type" : "object",
          "description" : "flags used by horton",
          "properties" : { }
        },
        "attributes" : {
          "type" : "object",
          "description" : "Message attributes",
          "properties" : { }
        }
      },
      "description" : "body for an invoming or outgoing event or message",
      "example" : {
        "horton_flags" : "{}",
        "attributes" : "{}",
        "body" : "{}"
      }
    },
    "methodRequestAndResponse" : {
      "type" : "object",
      "properties" : {
        "requestPayload" : {
          "type" : "object",
          "description" : "payload for the request that arrived from the service.  Used to verify that the correct request arrived.",
          "properties" : { }
        },
        "responsePayload" : {
          "type" : "object",
          "description" : "payload for the response to return to the service.  Used to verify that the correct request arrived.",
          "properties" : { }
        },
        "statusCode" : {
          "type" : "integer",
          "description" : "status code to return to the service"
        }
      },
      "description" : "parameters and response for a sync method call"
    },
    "methodInvoke" : {
      "type" : "object",
      "properties" : {
        "methodName" : {
          "type" : "string",
          "description" : "name of method to invoke"
        },
        "payload" : {
          "type" : "object",
          "description" : "payload to send to the method",
          "properties" : { }
        },
        "responseTimeoutInSeconds" : {
          "type" : "integer",
          "description" : "number of seconds to wait for a response"
        },
        "connectTimeoutInSeconds" : {
          "type" : "integer",
          "description" : "number of seconds to wait for the module to connect"
        }
      },
      "description" : "parameters used to invoke a method"
    }
  }
}